Added dependencies for Testcontainers, Testcontainers's junit jupiter, Testcontainers's postgreSQL. Added Test for CustomerJDBCDataAccessService class, updated `id` data type to long in the class.

+++ src/main/java/com/bachir/customer/CustomerDao.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
     void insertCustomer(Customer customer);
     boolean existsPersonWithEmail(String email);
-    boolean existsPersonWithId(Integer id);
-    void deleteCustomerById(Integer id);
+    boolean existsPersonWithId(long id);
+    void deleteCustomerById(long id);
     void updateCustomer(Customer customer);
 }
